+++ project/code/Common_Peripherals.h
@@ -46,6 +46,6 @@ void Buzzer_init();//蜂鸣器初始化
 void Key_init();//按键与LED初始化
 void key_scan();//按键扫描
 void Steer_init();
-void Steer_set(int angle);
+void Steer_set(float angle);
 
 #endif
